requirements.txt
gdelt==0.1.14
fredapi==0.5.2
sec-edgar-api==1.1.0
secedgar==0.6.0
sec-edgar-downloader==5.0.3
finnhub-python==2.4.25
python-dotenv==1.2.1
yfinance==0.2.66
# Pydantic
pydantic==2.12.5

# GCP
google-cloud-storage==2.18.2
google-auth==2.29.0

# Chroma
chroma

# Trading Agents
typing-extensions
langchain-openai
langchain-experimental
pandas
yfinance
praw
feedparser
stockstats
eodhd
langgraph
chromadb
setuptools
backtrader
akshare
tushare
finnhub-python
parsel
requests
tqdm
pytz
redis
chainlit
rich
questionary
langchain_anthropic
langchain-google-genai
# OpenAI Agents
openai-agents==0.6.2

# FastAPI
fastapi
uvicorn[standard]
httpx
nest-asyncio
pytest-asyncio
pytest-mock


# OpenAI Chat Application
openai
instructor

# Cloud SQL
pg8000
cloud-sql-python-connector[pg8000]    
psycopg[binary]
reportlab>=4.0.0